restructures github gateway directory
