+++ src/main/java/dev/coding/springboot/gateway/github/GithubRestGateway.java
@@ -4,6 +4,9 @@ import dev.coding.springboot.configuration.ServiceEndpointProperties;
 import dev.coding.springboot.configuration.ServiceEndpointProperties.ServiceEndpoint;
 import dev.coding.springboot.gateway.GetRequestObject;
 import dev.coding.springboot.gateway.RestGateway;
+import dev.coding.springboot.gateway.github.domain.GithubUser;
+import dev.coding.springboot.gateway.github.request.GetUserByUsernameRequestObject;
+import dev.coding.springboot.gateway.github.request.GetUsersRequestObject;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
